requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
celery==5.3.4
twilio==8.10.3
razorpay==1.3.0
requests==2.31.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
